商城测试结束
