+++ digital-service/src/main/java/com/digital/controller/CategoryController.java
@@ -47,82 +47,4 @@ public class CategoryController {
         return Result.success(getListCategoryVoArrayList);
     }
 
-    /**
-     * 添加商品类别
-     * @param categoryAddReq
-     * @return
-     */
-    @PostMapping("/category")
-    @AuthCheck(mustRole = UserConstant.ADMIN_ROLE)
-    public Result addCategory(@RequestBody CategoryAddReq categoryAddReq) {
-        if (categoryAddReq.getParentId() == null || categoryAddReq.getStatus() == null
-        || categoryAddReq.getName() == null) {
-            return Result.error(ResultErrorEnum.PARAM_IS_ERROR.getMessage());
-        }
-
-        if (!categoryAddReq.getStatus().equals(0) || categoryAddReq.getParentId() < 0) {
-            return Result.error(ResultErrorEnum.PARAM_IS_ERROR.getMessage());
-        }
-
-        Category category = new Category();
-        BeanUtils.copyProperties(categoryAddReq, category);
-
-        boolean save = categoryService.save(category);
-        if (!save) {
-            return Result.error(ResultErrorEnum.OPERATION_ERROR.getMessage());
-        }
-
-        return Result.success();
-    }
-
-    /**
-     * 删除商品类别
-     * @param id
-     * @return
-     */
-    @DeleteMapping("/category/{id}")
-    @AuthCheck(mustRole = UserConstant.ADMIN_ROLE)
-    public Result deleteCategoryById(@PathVariable Integer id) {
-
-        if (id == null || id <= 0) {
-            return Result.error(ResultErrorEnum.PARAM_IS_ERROR.getMessage());
-        }
-
-        boolean removeById = categoryService.removeById(id);
-        if (!removeById) {
-            return Result.error(ResultErrorEnum.OPERATION_ERROR.getMessage());
-        }
-
-        return Result.success();
-    }
-
-    /**
-     * 更新商品类别信息
-     * @param id
-     * @param categoryUpdateReq
-     * @return
-     */
-    @PutMapping("/category/{id}")
-    @AuthCheck(mustRole = UserConstant.ADMIN_ROLE)
-    public Result updateCategoryById(@PathVariable Integer id,
-                                     @RequestBody CategoryUpdateReq categoryUpdateReq) {
-
-        if (id == null || id <= 0) {
-            return Result.error(ResultErrorEnum.PARAM_IS_ERROR.getMessage());
-        }
-
-        QueryWrapper<Category> categoryQueryWrapper = new QueryWrapper<>();
-        categoryQueryWrapper.eq("id", id);
-
-        Category category = new Category();
-        BeanUtils.copyProperties(categoryUpdateReq, category);
-
-        boolean update = categoryService.update(category, categoryQueryWrapper);
-        if (!update) {
-            return Result.error(ResultErrorEnum.OPERATION_ERROR.getMessage());
-        }
-
-        return Result.success();
-    }
-
 }
