+++ digital-service/src/main/java/com/digital/controller/ProductController.java
@@ -148,79 +148,6 @@ public class ProductController {
         return Result.success(getProductVoPage);
     }
 
-    /**
-     * 添加商品，限管理员
-     * @param productAddReq
-     * @return
-     */
-    @PostMapping("/product")
-    @AuthCheck(mustRole = UserConstant.ADMIN_ROLE)
-    public Result addProduct(@RequestBody ProductAddReq productAddReq) {
-        if (productAddReq.getCategoryId() == null || productAddReq.getName() == null
-        || productAddReq.getSubtitle() == null || productAddReq.getMainImage() == null
-        || productAddReq.getSubImages() == null || productAddReq.getDetail() == null
-        || productAddReq.getPrice() == null || productAddReq.getStock() == null || productAddReq.getStatus() == null) {
-            return Result.error(ResultErrorEnum.PARAM_IS_ERROR.getMessage());
-        }
 
-        Product product = new Product();
-        BeanUtils.copyProperties(productAddReq, product);
 
-        boolean save = productService.save(product);
-        if (!save) {
-            return Result.error(ResultErrorEnum.OPERATION_ERROR.getMessage());
-        }
-
-        return Result.success();
-    }
-
-    /**
-     * 删除商品，限管理员
-     * @param id
-     * @return
-     */
-    @DeleteMapping("/product/{id}")
-    @AuthCheck(mustRole = UserConstant.ADMIN_ROLE)
-    public Result deleteProductById(@PathVariable Integer id) {
-
-        if (id == null || id <= 0) {
-            return Result.error(ResultErrorEnum.PARAM_IS_ERROR.getMessage());
-        }
-
-        boolean removeById = productService.removeById(id);
-        if (!removeById) {
-            return Result.error(ResultErrorEnum.OPERATION_ERROR.getMessage());
-        }
-
-        return Result.success();
-    }
-
-    /**
-     * 更新商品信息，限管理员
-     * @param id
-     * @param productUpdateReq
-     * @return
-     */
-    @PutMapping("/product/{id}")
-    @AuthCheck(mustRole = UserConstant.ADMIN_ROLE)
-    public Result updateProductById(@PathVariable Integer id,
-                                     @RequestBody ProductUpdateReq productUpdateReq) {
-
-        if (id == null || id <= 0) {
-            return Result.error(ResultErrorEnum.PARAM_IS_ERROR.getMessage());
-        }
-
-        QueryWrapper<Product> productQueryWrapper = new QueryWrapper<>();
-        productQueryWrapper.eq("id", id);
-
-        Product product = new Product();
-        BeanUtils.copyProperties(productUpdateReq, product);
-
-        boolean update = productService.update(product, productQueryWrapper);
-        if (!update) {
-            return Result.error(ResultErrorEnum.OPERATION_ERROR.getMessage());
-        }
-
-        return Result.success();
-    }
 }
